pyproject.toml
[tool.poetry]
name = "ecommerce"
version = "0.1.0"
description = ""
authors = ["Adam Szustak <madasz@onet.eu>"]

[tool.poetry.dependencies]
python = "^3.8"
django = "^3.1"
Pillow = "^7.2.0"
psycopg2 = "^2.8.5"
django-allauth = "^0.42.0"
django-crispy-forms = "^1.9.2"
faker = "^4.1.2"
easy_thumbnails = "^2.7"
django-mptt = "^0.11.0"
django-mptt-urls = "^2.0.4"
django-localflavor = "^3.0.1"
celery = "^5.0.1"
braintree = "^4.4.0"

[tool.poetry.dev-dependencies]
django-debug-toolbar = "^2.2"
pre-commit = "^2.6.0"
black = "^19.10b0"
flake8 = "^3.8.3"
pytest-cov = "^2.10.0"
pytest-pythonpath = "^0.7.3"
factory-boy = "^3.0.1"
pytest-django = "^3.9.0"
django-ckeditor = "^5.9.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
